refactor : 패키지 이름 변경
+ 패키지 이름을 service -> model로 변경함
